backend/requirements.txt
fastapi
uvicorn
torch
openai-clip
torchvision
p
faiss-cpu
pillow
